+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## HEAD
 
+* Fix margins around modal in resize plugin
+
 * Fix position of close button when modal is between mobile and table
 * Avoid incorrect rendering of the detail container
 * Correct resizing and positioning in IE8